\name{covKernel-class}
\Rdversion{1.1}
\docType{class}
\alias{covKernel-class}

\title{Class "covKernel" }
\description{ Union of classes including "covTensorProduct", 
"covIso", "covAffineScaling", "covScaling" and "covUser"}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Methods}{
No methods defined with class "covKernel" in the signature.
}

\author{ Olivier Roustant, David Ginsbourger, Yves Deville }

\examples{
showClass("covKernel")
}
\keyword{classes}
