\name{drop.response}
\alias{drop.response}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Trend model formula operation }
\description{  Drop the response in the formula specifying the linear trend. }
\usage{
drop.response(formula, data)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{formula}{  an object of class formula. }
  \item{data}{ a data frame corresponding to formula.}
}

\value{ An object of class formula. }

\author{ O. Roustant, D. Ginsbourger }

\seealso{ \code{\link{trendMatrix.update}} }

\keyword{models}
\keyword{regression}
\keyword{internal}
