\name{inputnames}
\alias{inputnames}


\title{Get the input variables names}

\description{Get the names of the input variables.}

\usage{
  inputnames(x)
}

\arguments{
 \item{x}{ an object containing the covariance structure.}
}

\value{
  A vector of character strings containing the names of the input variables.
}

\keyword{models}
