\name{kernelname}
\alias{kernelname}


\title{Get the kernel name}

\description{Get the name of the underlying tensor-product covariance structure.}

\usage{
  kernelname(x)
}

\arguments{
 \item{x}{ an object containing the covariance structure.}
}

\value{
  A character string.
}

\keyword{models}
