\name{ninput}
\alias{ninput}


\title{Get the spatial dimension}

\description{Get the spatial dimension (number of input variables).}

\usage{
  ninput(x)
}

\arguments{
 \item{x}{ an object containing the covariance structure.}
}

\value{
  An integer equal to the spatial dimension.
}

\keyword{models}
