\name{nuggetvalue}
\alias{nuggetvalue}
\alias{nuggetvalue<-}

\title{Get or set the nugget value}

\description{Get or set the nugget value.}

\usage{
  nuggetvalue(x)
  nuggetvalue(x) <- value 
}

\arguments{
  \item{x}{ an object containing the covariance structure.}
  \item{value}{ an optional variance value standing for the homogeneous nugget effect.}
}


\keyword{models}
